+++ Engine/Source/Renderer/Material/Material.cpp
@@ -1,6 +1,7 @@
 #include "Material.h"
 #include "Base.h"
 #include "Core/Logger.h"
+#include "Renderer/Material/MaterialSystem.h"
 #include "Renderer/Texture/TextureSystem.h"
 
 namespace Core
@@ -21,32 +22,44 @@ namespace Core
 
     Material::~Material() { ClearTexture(colorTexture); }
 
-    void Material::ClearTexture(Texture2D *texture)
+    void Material::ClearTexture(TexPair &p)
     {
-        if (colorTextureType == TextureType::None)
+        if (p.Type == TextureType::None)
         {
             CE_LOG("CE_RENDER", Error, "Texture type cannot be None.");
             return;
         }
 
-        if (colorTextureType == TextureType::Default)
+        if (p.Type == TextureType::Default)
             return;
 
-        TextureSystem::Remove(colorTexture->GetImagePath());
-        colorTexture = nullptr;
-        colorTextureType = TextureType::None;
+        TextureSystem::Remove(p.Tex->GetImagePath());
+        colorTexture.Tex = nullptr;
+        colorTexture.Type = TextureType::None;
     }
 
     void Material::SetColorTextureDefault()
     {
-        colorTextureType = TextureType::Default;
-        colorTexture = TextureSystem::GetDefault();
+        if (colorTexture.Tex)
+            ClearTexture(colorTexture);
+
+        colorTexture.Type = TextureType::Default;
+        colorTexture.Tex = TextureSystem::GetDefault();
     }
 
     void Material::SetColorTexture(const std::string &name)
     {
-        colorTextureType = TextureType::Loaded;
-        colorTexture = TextureSystem::Get(name);
+        ClearTexture(colorTexture);
+
+        if (this == MaterialSystem::GetDefault())
+        {
+            CE_LOG("CE_RENDER", Warn,
+                   "Cannot set color texture, material must be set to Config type");
+            return;
+        }
+
+        colorTexture.Type = TextureType::Loaded;
+        colorTexture.Tex = TextureSystem::Get(name);
     }
 
     void Material::Use(Shader *shader)
@@ -55,9 +68,9 @@ namespace Core
 
         shader->Vec4(state.Color, "uColor");
 
-        CE_VERIFY(colorTexture);
-        colorTexture->Use();
+        // Due to per instance draw (as bad as it is) generations are really needed
+        CE_VERIFY(colorTexture.Tex);
+        colorTexture.Tex->Use();
         shader->Int(0, "uColorTexture");
     }
-
 } // namespace Core