+++ Engine/Source/Renderer/Object/Mesh.h
@@ -1,7 +1,9 @@
 #pragma once
 
 #include "Base.h"
+#include "Math/Vector.h"
 #include "Renderer/Buffer/VertexArray.h"
+#include "Renderer/Geometry/Geometry.h"
 #include "Renderer/Material/Material.h"
 #include "Renderer/Shader/Shader.h"
 
@@ -24,20 +26,25 @@ namespace Core
     private:
         VertexArray *geometryArray;
         Material *material;
+        Geometry *geometry;
 
         void DestroyGeometryArray();
         void BufferGeometryArray();
 
         void DestroyMaterial();
+        void DestroyGeometry();
 
     public:
+        Vector3 Position;
+
         Mesh();
         ~Mesh();
 
         void SetMaterialToDefault();
-
         void SetMaterial(const std::string &name);
         void SetMaterial(const Material::Configuration &config);
+
+        void SetGeometry(Geometry *geometry);
 
         inline Material *GetMaterial() { return material; };